+++ cdk/lib/main-function-subsconstruct.ts
@@ -70,6 +70,7 @@ export class AutoAlarm extends Construct {
       'AutoAlarm-RdsCluster',
       'AutoAlarm-Route53resolver',
       'AutoAlarm-Sqs',
+      'AutoAlarm-Sfn',
       'AutoAlarm-TargetGroup',
       'AutoAlarm-TransitGateway',
       'AutoAlarm-Vpn',
@@ -282,6 +283,15 @@ export class AutoAlarm extends Construct {
       new PolicyStatement({
         effect: Effect.ALLOW,
         actions: ['ec2:DescribeVpnConnections'],
+        resources: ['*'],
+      }),
+    );
+
+    // Attach policies for Step Functions to list statemachine and get tags
+    autoAlarmExecutionRole.addToPolicy(
+      new PolicyStatement({
+        effect: Effect.ALLOW,
+        actions: ['states:ListStateMachines', 'states:ListTagsForResource'],
         resources: ['*'],
       }),
     );